requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
et-xmlfile==1.1.0
filelock==3.12.4
Flask==2.2.5
Flask-Migrate==3.0.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
gTTS==2.3.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
llvmlite==0.41.0
Mako==1.2.4
MarkupSafe==2.1.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.58.0
numpy==1.22
openai==0.28.0
openai-whisper==20230918
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2023.8.8
requests==2.27.0
six==1.16.0
SQLAlchemy==2.0.21
sympy==1.12
tiktoken==0.3.3
toml==0.10.2
torch==2.0.1
tqdm==4.62.3
typing_extensions==4.8.0
urllib3==1.26.7
Werkzeug==2.2.2
yarl==1.9.2
